Changed fusion structure to be included in cards
Cards now carry the product and the type of material (1 or 2) that they are meant to be. This should theoretically result in faster times when checking if two cards can fuse.

External fusions (called `externals` in the schema) are to be used only when a set is dependent on another set, and will use the `fusions` array, by referring to the index of the fusion in the array.

The point of externals is that, in case someone wants to make an add-on set to another set, they can still be compatible with one another, by simply referring to their parent set only when necessary. Thus, the `set` string should only be used to check a parent set. This should theoretically also allow for retroactive compatibility with grandparent sets (add-on to an add-on).

The expected behavior of fusions now should be:
- Check if both cards have fusions.
- Compare if both cards share a product, and are a different material.
- If there are no other fusions in either array, check for externals, and check them if they exist.
- Otherwise, discard the first card.