Implemented turn end
The player can now end their turn by pressing the Enter (Start) button. This moves the cursor, as well as rotates the screen 180 degrees, while also reversing the controls, so that the white player can play normally.

During the turn ending, the cursor vanishes, while a tile indicator is used to slowly disappear in place of the cursor, which is the parent scene of the camera, in order to maintain the illusion of the cursor remaining in place.