Minor edit to just_spellbound
The `just_spellbound` flag now flips down for ALL cards at the end of the turn.

In the original game, if a card is spellbound on its owner's turn, the spellbound counter will not decrement. However, if it's spellbound on its opponent's turn instead, then it will decrement at the end of its owner's turn.