Added basic hand functionality
The game now opens a hand that draws up to five cards from the top of the deck, which is shuffled at the start of the game.

A queue is used for preparing the order cards should be played, and works near identically to how it did in the original game.

The next step should be actually adding fusions into the game, and adding the result from any action in the hand onto the field, which should put the game into a half-playable state.