Most hand animations done, card data transfers
Cards in the hand now animate to prepare for a fusion, as well as when reversing. All that's left now is prepare the fusion animations.

Cards can now copy and paste data from one to another. This will be useful for transferring the final card to the field, after fusions and power ups are applied.