Minor adjustments, bug fixes to hand and cards
The hand can now be opened post-summon to show the current hand, without allowing summoning, acting as close as possible to the original game.

Hand card moving to field animation also fixed, by changing its parent from the HUD to the board, so that it remains in the intended location.