Amended Effects in Database
Effects are now split in three parts:
- Trigger: A check that activates the effect
- Effect: Something that affects the game or field
- Target: What gets affected by the Effect

All three can use attributes, depending on context. In-engine processing of effects will occur by checking these attributes, along with the type of Trigger, Effect, and Target selected.

There can be multiple effects in a single card, and each effect may have multiple Triggers and Targets, which work as AND statements. For example, a Target being `monsters` and `row_current` will affect any Monster on the row the card which activated the effect is currently on.