Basic effect implementation + schema edits
When a trigger check is to be made, the process_trigger function is called. This checks if a trigger is valid and can activate the effects. This now allows a basic framework on which to build effects as intended, while also allowing skipping the trigger step for spells and limited range traps.

From this point on, an effect may only have one trigger. This is to be closer to the original game, as well as make the process easier to code. This is reflected in the schema and test database.

Another debug button was added, to test effects and terrain.