Limited traps now works as intended
The check for adjacent cards happens at the end of a card's movement to another tile, so a card moving to two tiles will have that check happen twice in total.

The oldest enemy limited trap will activate, as per the original game's rules, and if coded properly, should destroy itself before the effects begin resolving.

Also, spellbind counters are now implemented, and will appear next to the ATK/DEF text if face-up, or at the center of the card if face_down. This is a workaround to the inability to invert colors to a card in Godot 3, and the current lack of a GUI.