Another trigger and target, field fusions
The "flipped_face_up_battle" trigger has been implemented. This triggers when a battle is about to start, and the card gets flipped. The attacking card is flipped first, and the defending one last. Then, battle calculation begins, provided both cards survived the effects.

The "opponent" target has also been implemented. Considering that this only applies during battles, this target will fail in any other context.

Field fusions also function, using the same function as hand fusions, to determine the result. In case of a failed fusion, the second card is destroyed, and the first one moves to the second one's location. A placeholder for power-up cards has been added for later.