Implemented Deck Editor
The editor uses two lists, one for sets, and one for the two player decks. The player activates an option from the sets and inputs a card if possible. When activating an option from the deck, that card is removed. The player will be able to start a Duel once both decks have 40 cards each.

Note: Due to Godot 3.5 not allowing for the ability to disable incremental keyboard search on ItemLists, buttons for changing the sets and decks have been provided for convenience.

Assuming no game breaking bugs exist, this prototype can be considered complete for presentation!